Refactored tests (corrected integration tests vs normal tests)
