+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/column/AggregationFunctionsIT.scala
@@ -1,6 +1,7 @@
 package com.crobox.clickhouse.dsl.column
 
 import com.crobox.clickhouse.DslITSpec
+import com.crobox.clickhouse.DslITSpec.StringResult
 import com.crobox.clickhouse.dsl._
 import spray.json.DefaultJsonProtocol._
 import spray.json.RootJsonFormat
