+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/column/EncodingFunctionsIT.scala
@@ -1,11 +1,11 @@
 package com.crobox.clickhouse.dsl.column
 
-import com.crobox.clickhouse.DslIntegrationSpec
+import com.crobox.clickhouse.DslITSpec
 import com.crobox.clickhouse.dsl._
 
 import java.util.UUID
 
-class EncodingFunctionsTest extends DslIntegrationSpec {
+class EncodingFunctionsIT extends DslITSpec {
 
   it should "succeed for EncodingFunctions" in {
     r(hex(12)) shouldBe "0C"
