+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/column/HashFunctionsIT.scala
@@ -1,9 +1,9 @@
 package com.crobox.clickhouse.dsl.column
 
-import com.crobox.clickhouse.DslIntegrationSpec
+import com.crobox.clickhouse.DslITSpec
 import com.crobox.clickhouse.dsl._
 
-class HashFunctionsTest extends DslIntegrationSpec {
+class HashFunctionsIT extends DslITSpec {
 
   it should "succeed for HashFunctions" in {
     val someStringData = "fooBarBaz"
