+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/column/HigherOrderFunctionsIT.scala
@@ -1,9 +1,9 @@
 package com.crobox.clickhouse.dsl.column
 
-import com.crobox.clickhouse.DslIntegrationSpec
+import com.crobox.clickhouse.DslITSpec
 import com.crobox.clickhouse.dsl._
 
-class HigherOrderFunctionsTest extends DslIntegrationSpec {
+class HigherOrderFunctionsIT extends DslITSpec {
   val arr1 = Seq(1L, 2L, 3L)
 
   it should "HigherOrderFunctions: arrayAll" in {
