+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/column/URLFunctionsIT.scala
@@ -1,9 +1,9 @@
 package com.crobox.clickhouse.dsl.column
 
-import com.crobox.clickhouse.DslIntegrationSpec
+import com.crobox.clickhouse.DslITSpec
 import com.crobox.clickhouse.dsl._
 
-class URLFunctionsTest extends DslIntegrationSpec {
+class URLFunctionsIT extends DslITSpec {
 
   it should "succeed for URLFunctions" in {
     val someUrl        = "https://www.lib.crobox.com/clickhouse/dsl/home.html?search=true#123"
