+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/column/UUIDFunctionsIT.scala
@@ -1,5 +1,6 @@
 package com.crobox.clickhouse.dsl.column
 
+import com.crobox.clickhouse.DslITSpec.StringResult
 import com.crobox.clickhouse.dsl._
 import com.crobox.clickhouse.{dsl, DslITSpec}
