+++ dsl/src/it/scala/com/crobox/clickhouse/dsl/language/CollectionFunctionTest.scala
@@ -1,8 +1,9 @@
 package com.crobox.clickhouse.dsl.language
-import com.crobox.clickhouse.DslIntegrationSpec
+
+import com.crobox.clickhouse.DslITSpec
 import com.crobox.clickhouse.dsl._
 
-class CollectionFunctionTest extends DslIntegrationSpec {
+class CollectionFunctionTest extends DslITSpec {
 
   it should "succeed for IN functions" in {
     val someCollection = Seq(1, 4, 6, 9)
